tests/wpt/web-platform-tests/websockets/interfaces/CloseEvent/clean-close.html

<!doctype html>
<title>WebSockets: wasClean, true</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onopen = t.step_func(function(e) {
    ws.send('Test');
  });
  ws.onmessage = t.step_func(function(e) {
    ws.close();
  });
  ws.onclose = t.step_func(function(e) {
    assert_equals(e.wasClean,true);
    t.done();
  });
}, null, {timeout:2000});
</script>
